Rename slice_copy to slice_owned
