+++ src/impls.rs
@@ -9,7 +9,7 @@ pub mod offsets;
 pub mod option;
 pub mod result;
 pub mod slice;
-pub mod slice_copy;
+pub mod slice_owned;
 pub mod storage;
 pub mod string;
 pub mod tuple;
