+++ src/impls/slice_owned.rs
